requirements.txt
pydantic>=2.5.2
chromadb>=0.4.17
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
numpy>=1.26.2
pandas>=2.1.3
sqlalchemy>=2.0.23
aiohttp>=3.9.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
tenacity>=8.2.3
loguru>=0.7.2
langchain>=0.0.350
langchain-community>=0.0.10
sentence-transformers>=2.2.2
telethon==1.34.0
# Deep Learning
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.1+cu121
